feat: add eventListener on buyNow button
